`om init`: Switch to lockless registry
The registry is now defined in a `registry.json` file, much like flake registry. This gets rid of the exponentially increasing `flake.lock` file, which impacts `om init`'s initial loading performance. The JSON is embedded in the Rust binary, leading to faster load times (since we only need to parse it at runtime, once).

Minor downside: the tests are not 100% reproducible anymore, since they will be using the latest HEAD of these repos. But that may be desired in that it forces us to test `om init` against the latest templates.