Fix: `just cargo-test` was erroneously running `cargo doc` (#356)
Also fixes tests failing due to deprecation of `nixci` and `nix-health`
flake configs. See

https://github.com/juspay/omnix/blob/df0093881be956a9df8034e5970dc363838bf478/doc/src/history.md?plain=1#L33


Notes on changes:
- `test_services_flake` doesn’t run `om ci` on all the sub-flakes (which
was the case before) because unlike before we have a lot of services and
since the goal of the test is to just check if the `stdout` is not
polluted with non-outPaths, testing a smaller sub-flake should suffice.
- `test_config_loading` upgrades the commit of `haskell-flake` to the
one that uses `om.ci` configuration.