ci: Remove 'Added input' filtering. (#437)
And pass STDERR as is to the user's terminal.

---

I don't think the complexity here is worth it for the slight benefit of
keeping `om ci run`'s output less verbose. The ideal solution here is to
fix upstream Nix to not spew out these messages if the user passes
`--quiet --no-write-lock-file`. See here where it unconditionally writes
this:


https://github.com/NixOS/nix/blob/8e8edb5bf857d62f7295c15534d2a4e555065fdf/src/libflake/flake/flake.cc#L906-L909

This change incidentally resolves #428