Release omnix-cli 1.3.1
Fixes the `versionCheckHook` failure in nixpkgs. See error logs:
https://github.com/NixOS/nixpkgs/pull/472294#issuecomment-3673963729

This should've been done in #490, but `cargo workspaces publish` only
updated the crates that changed since the last release. Since
`omnix-cli` is the public interface for all the crates, it makes sense
to always bump it, hence I have updated the README to include `--force
omnix-cli` while publishing a release.